+++ src/Core/AuthorizationException.php
@@ -2,6 +2,12 @@
 
 namespace Niang\Core;
 
-class AuthorizationException extends \RuntimeException
+use Niang\Core\Exceptions\HttpException;
+
+class AuthorizationException extends HttpException
 {
+    public function __construct(string $message = '')
+    {
+        parent::__construct(403, $message !== '' ? $message : 'Action non autorisée.');
+    }
 }