+++ src/Core/Exceptions/AuthenticationException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Niang\Core\Exceptions;
+
+class AuthenticationException extends HttpException
+{
+    public function __construct(string $message = '')
+    {
+        parent::__construct(401, $message);
+    }
+}